+++ common/Makefile
@@ -1,5 +1,6 @@
 COMMON_SRC_NOGEN :=				\
 	common/bech32.c				\
+	common/bech32_util.c			\
 	common/bip32.c				\
 	common/bolt11.c				\
 	common/channel_config.c			\
