+++ devtools/Makefile
@@ -5,6 +5,7 @@ DEVTOOLS_TOOL_OBJS := $(DEVTOOLS_TOOL_SRC:.c=.o)
 
 DEVTOOLS_COMMON_OBJS :=				\
 	common/bech32.o				\
+	common/bech32_util.o			\
 	common/bolt11.o				\
 	common/hash_u5.o			\
 	common/type_to_string.o			\
